Make the Content Performance Manager absent
This application has been migrated to AWS under the Content Data API
name, so it can now be removed from the Carrenza environments, and
Integration.

This commit should get govuk-puppet to remove some parts of the
application, and then once these changes are deployed, the
configuration within govuk-puppet can be removed.