Add sidekiq-monitoring to Production AWS and Staging AWS
We're running applications using Sidekiq in these environments, so it
makes sense to run sidekiq-monitoring here as well.

It's not going to work very well, as it's not aware what's running in
Carrenza, and what's running in AWS, but this might be more useful
than not running it in AWS entirely.