allow defined logrotate to lookup hiera
if the defined logrotate is able to lookup hiera, it is possible to modify the configuration in the hiera rather than pass the value at instanciation of the defined resource. This is a hack since the class and other defined types of logrotate should have exposed the logrotate parameters for configuration.