lower puppet memory usage
Before, it was thought that giving more memory to puppetserver will solve the out-of-memory issue but it didn't. Now we use monit to restart the puppetserver if it is out-of-memory and stops. Hence, we can lower the memory allocated to puppetserver and avoid to be within 5% of memory usage on the virtual machine.