Configure the environment sync for the Content Data Admin
The Content Data Admin does use a PostgreSQL database, so for
consistency, sync it between environments. It doesn't contain any data
other than user data stored by gds-sso, so this isn't very useful.

Some of the configuration was already in place, so just put the rest
in, and enable the absent tasks.

Also change the integration pull task to pull from Production, rather
than Staging, as this is how the data syncs have worked in the past,
and I couldn't find any reason to do differently.