Restore Icinga Slack notifications
These may be useful in making people more aware of when alerts occur.

I'm also looking at this as a way of making alerts more visible to
specific teams, but I want to get something rough and general working
first.

This has worked in the past, but I think probably broke when the Slack
instance changed, or the Slack API changed.