Remove redundant Content Performance Manager configuration
This application has now been renamed to the Content Data API. Various
things have been set to 'absent', and these changes have been
deployed, so now the Puppet configuration can be removed.

There's still some work to do on the development-vm, and the now
redundant nodes, but this'll come later.