Remove some confusing node_class hieradata
From AWS Staging and AWS Production. This configuration suggests that
there are publishing_api, email_alert_api, whitehall_backend and
whitehall_frontend machines in these environments, and that we're
running applications on them.

This is confusing, as these machines do not exist, so the applications
cannot be running on them.

As I don't know when this will likely change, lets remove the
configuration so that the hieradata overall is simpler.