Do more testing on the NGinx configuration
Currently, it's possible for govuk-puppet to set NGinx up with invalid
configuration, but this not to cause govuk-puppet to always fail. I'm
guessing that govuk-puppet may fail when it restarts the service, but
then succeeds on subsequent runs.

NGinx is important, and testing the configuration doesn't take a long
amount of time, so do this on every run. Ideally there'd be a neater
way of doing this with Puppet, but I don't know of one.