Remove ckan load balancing from the backend_lb class
I believe CKAN is now hosted in AWS, which doesn't use this node type,
so this configuration is now redundant. This follows on from [1],
which removed some hieradata that this functionality was dependant on.

1: b12ea3fbcca6af3cdbca22236d59ebe9bf600940

With that change [1], the generated NGinx configuration is invalid, as
there aren't any upstream servers for CKAN defined.