Update to the 2.1.1 release of icinga-slack-webbook
This includes improved output and display for the Slack messages. The
Slack Icinga contact configuration is updated accordingly.

The $HOSTNAME$ as well as the $HOSTDISPLAYNAME$ is passed to the
icinga_slack_webhook_notify script, so that it can construct the URL
to the extinfo.cgi page for a service alert.