Add LB healthcheck redirection apps
`kibana` and `content_data_admin` apps implement the `nginx::config::vhost::redirect`
class instead of the standard proxy one. The LB healthchecks that we configured in 1f6eec5
do not cover this special cases, and we need to add the application healthcheck endpoint
to make sure the app gets traffic from the loadbalancer.