Fix kibana, content-data-admin lb healthcheck resources
Fix 0c83d1c

The healthcheck configurations are not being created,
`defined(Concat['/etc/nginx/lb_healthchecks.conf'])` is not evaluating as `true`.
It might be evaluated before the resource is created in the s_backend manifest
(the function depends on the order parse of the config).

It should be OK to create the concat fragments in any case, so we are just
going to evaluate if we are on the Integration environment and remove the
conditional when we need to promote to the rest of environments.

Other fixes are the need of `"` to parse `\n` correctly, and avoid a duplicated
`concat::fragment` resource name in `content_data_admin.pp`.