Fix asset slave scripts not reporting errors
Previously we (re)set the nagios code to report based on the last
command run by the script, which was often not the command that failed
(e.g. an 'echo' to print out an error message). This changes the way we
set the exit status to ensure the status is assigned correctly on error.