Enable LB app health checks on Staging
Deploy the application health check feature on Staging.

On Integration we had Nginx configtest errors temporary if
`/etc/nginx/lb_healthchecks.conf` wasn't created before the nginx conf
reloaded with the `include` statement. We are going to try to force
the `concat` resource to be created before the Nginx default vhost.