Add Cache ALB health check path
Define location paths for application health checks from the public ALB.

We want to add the instance to the ALB `www` target group when Nginx, Varnish and
Router are up. We can achieve this with the `__canary__` endpoint. This endpoint
is not cached in Varnish.

`assets` traffic will be available as soon as Nginx is healthy.