Use write_graphite instead of network plugin
Previously we used the collectd network plugin in a client/server
configuration where all instances sent their collectd stats to the
graphite server, which forwarded them to graphite itself. Using the
network plugin is inconsistent between Carrenza and AWS. This changes
the collectd config on all machines to use the graphite plugin
exclusively, which does not seem to need additional firewall config.