Overrides the process regex to look for custom tag
We've had to pass a custom '--tag' to sidekiq so that it names our
process. Previously, process would simply be 'sidekiq 5.2.5 content-audit-tool [0 of x]',
which doesn't tell us which of the workers it is.

Now, output will be 'sidekiq 5.2.5 content-audit-tool/google-analytics-worker [0 of x]'
(for example), which gives us a very regexable process name.