Increase memory monitoring thresholds for licensify
The CollectD process plugin has only just been configured correctly,
so this is something that's only now appearing.

Levels seem to be around the 600 MB - 800 MB mark for the past two days.
Think it's sensible to set a warning at 900 MB and a critical warning
at 1000 MB. We can revisit these later if things change.