Update package py-dateutil resource in collectd docker
On the ci Jenkins machines, Puppet keeps reinstalling the package
py-dateutil on every run. It might be that it's not able to determine
the current version vs the version required. There are a few bugs in
the package pip provider, might be a proble of running an old version of
Puppet.

In other cases, we enforce the version of the pip package with the
`ensure` parameter, instead of the resource name. We can try this
to see if we stop the reinstallations.