Fix smokey test issues
Previously we installed a system version of chromedriver, which
conflicted with our use of the webdrivers gem. This stops puppet trying
to manage chromedriver and lets webdrivers do it instead.

Previously we iterated the upstart config for the smokey-loop service to
use 'setuid' instead of 'sudo -u', which doesn't seem to work. This
reverts the use of 'setuid' back to 'sudo -u' and enables a login shell
for the smokey user to better facilitate debugging.