remove carrenza mirror and update pingdom after mirror redesign
Following the mirror redesign, we will no longer need the carrenza mirrors and therefore we stop syncing the mirrorer with the carrenza mirrors. As a consequence, we also stop monitoring the mirrors via pingdom.

Furthermore, we start monitoring the s3 buckets. Right now we can't monitor GCS bucket due to Pingdom plan restrictions.