Add the Content Data apps back to the Integration Deploy Jenkins job
These were removed (e.g. [1] from the `deployable_applications` list
in the Carrenza side hieradata, as they're no longer running in
Carrenza.

1: 5c729768f6277d9ad716d3faa411a3c19fa5afd9

However, they still need to be present in the Integration Deploy job,
as this is actually deploying in the Integration environment running
in AWS.

This should resolve the failing master branch builds for these
applications.