Replace transition load all data with two specific jobs
This introduces two jobs: transition_import_hits and
transition_import_dns that will run daily to replace the
transition_load_all_data job. The reason this is replaced is because the
load_all_data job was somewhat strange in that it ran a job nightly that
could update fastly CDN which was configured as a separate job to only
run when the transition config repo changed.

This has meant that every day (for years) we have been updating the
Fastly CDN out of hours to essentially no effect. This was until it
broke and called us up at 4am.

The load_all_data job performed 3 tasks [1] which were: update sites
from config, import hits, and update DNS records. The first of these
tasks is covered by the other transition_load_site_config job that only
runs in work hours on weekdays (on a poll of the transition-config repo
for changes). The second is done be the import_hits job here, which will
run at 3am daily like the previous incarnation. The third, update DNS
records, is covered by the import_dns job which will run daily at 12
since there doesn't seem to be any special reason for it to run out of
hours.

The old load_all_data job has now been marked as absent to be removed
from the machines. Once deployed this can be removed from this repo.

As part of this change I've simplified the HOSTNAME lookup, since these
jobs only run on AWS and the hostname lookup would work on carrenza
anyway. I've removed the cloning of transition-config from the jobs as
this is not needed if sites are not being imported (I hope). I've
removed the downstream Bouncer_CDN job as that should only be called
when sites are transitioned. Finally, I've removed
enable_slack_notifications for these jobs since it is not an available
parameter so this setting had no effect.

[1]: https://github.com/alphagov/transition/blob/76aab40ea21ce44426f2335ef4034b55514ad2d2/lib/tasks/import/all.rake#L7-L11