Remove bouncer/transition jobs from Staging jenkins
When I restored these following [1] I was under the impression that
these jobs used to always run in staging and only recently had been
broken with the redirects. It turns out I was wrong in this assertion
and that they'd only been added to staging in Oct 2018 [2].

Having a had a couple of chats with people it doesn't seem ideal to have
them run in staging as the Bouncer CDN job is at risk of overwriting
production and the other scheduled jobs won't really offer any early
warnings. Therefore it seemed simpler to return these to only return in
production. This is a shame, as it's nice to have as little as possible
only run in production but I'm just not sure what the risks are of
having bits of this run differently.

This also fixes an alphabetical issue with the job list.

[1]: https://github.com/alphagov/govuk-puppet/commit/8780523b0123c0ceb59a502300158f2b1827b088
[2]: https://github.com/alphagov/govuk-puppet/commit/e985e342004881d0907d832dd7d5456465a8ba50