Use default time on response times panel
The response times panel is used on the [search-api dashboard](https://grafana.blue.staging.govuk.digital/dashboard/file/search-api.json).

The dashboard has a default time set to "Last 1 hour", which all the panels
honour except this one, because it has an override time set.

I suspect this is just an oversight because it's not used on the "standard"
rails dashboards.