Handle deploying govuk-content-schemas as part of deploy all apps
GOV.UK Content Schemas is a somewhat special example where we have code
that is deployed to GOV.UK infrastructure but it isn't an app that runs
so it lacks a puppet manifest. This code is deployed by the deploy app
Jenkins job and is something that needs to occur when a node is built
for the first time.

The code that defines which apps run on a node class is the same as that
which is used to define which projects are deployed. This makes it
problematic to define that content schemas should be deployed to
certain nodes, without having it also treated as an app that should
run on the node.

To work around this I set up the concept of non_apps that is part of the
definition for projects on a particular node. I'm not particular pleased
with myself for that name and anticipate it may raise a WTF, I like to
think though that it's the concept of the project that is a bit strange
and that the syntax for defining it consistent with the oddness that is
already present.

This was added after the node instances were destroyed in integration
and when re-built the Publishing API was no longer operating correctly.