Fix bouncer_cdn job_url
It's currently only configured in Carrenza so you get a 404 link if you
try and click on the link in Icinga.

This commit changes the URL to be hard coded to the AWS URL as there
currently isn't an easy way of constructing that URL programatically.

I've based this on what was done for
https://github.com/alphagov/govuk-puppet/pull/8935.