Stop listing migrated applications as deployable
In the Carrenza specific hieradata. These applications have been
migrated to AWS, and therefore aren't deployable in the Carrenza sides
of Production or Staging, and also aren't deployable in the CI
environment.

Instead, just include them specifically where the configuration is
needed. For the CI integration_deploy job, and the CI pipeline jobs.