disable active mirror crawling AWS integration
we switch off active mirror crawling in AWS integration to save money.
We still keeps the instance because it is needed in the deployment pipeline of govuk_crawler_worker.

govuk-crawler-worker must be installed on the machine to allow Jenkin deploys to work in integration.
Moreover, ffi gem has to be installed to a specific version because otherwise govuk-crawler wants to
install a ffi version which requires ruby >= 2.0