Point draft frontend apps at live search
We currently don't run a search instance in the draft stack which means
that users see internal server errors when they navigate to pages which
require search.

There is a plan to eventually run a draft search instance
https://github.com/alphagov/govuk-rfcs/blob/master/rfc-086-draft-stack-rummager.md
but for now as we don't have that, it's better that users see out of
date information rather than an error page.