Avoid false 429 rate alerts whenever a new router instance comes up.
When a new `router` instance appears, the '429 rate' alert fires with UNKNOWN status.
This happens because the statsd metric doesn't get created until the first 429 response
happens (which doesn't and shouldn't normally happen), which leads to the check_graphite
command failing because it can't find the timeseries, leading to the UNKNOWN status and
alerts appearing on the Icinga dashboard.

`check_graphite` has an `--ignore_missing` option which does exactly what we want in
this particular case. That is, we want the check to succeed if there is no metric for
429 rate on a given router instance.

The `--ignore-missing` option was added in a more recent version of `check_graphite`
than the one presently installed, so we also update the version here.