Read mongo collections using printjson
The format of db.getCollectionNames() has changed between versions of
Mongo. In Mongo 2 you get back a string delimited by a comma but in
Mongo 3 you get a full JSON array as a string. By using `printjson` we
can make the format the same on both versions.

I've tested this on Mongo version 3.2.7 and 2.4.9.