with_reboot_lock script
we would like to use our locksmith based lock holding script wrapper in
place of setlock so that we can use the same locking logic to prevent
reboots during env-sync and friends.

add a script that holds onto the distributed reboot lock for the
machine-type.

we have seen that setlock is being used in a way that is isn't quite
designed to (expected that a lock file gets deleted after free - when it
actually uses an exclusive write lock on the file not the file itself).

to avoid these issues we intend to replace uses of setlock that are
intending to hold the reboot lock with this script.