make docker_management not safe_to_reboot
The docker_management machine holds the reboot mutex/locks for the other
machines (in etcd) and it is not deployed in a HA configuration. This
means it is not safe to reboot this machine while other machines are
holding reboot locks.

Rather than running a HA etcd (which would be a pain) we exclude this
machine from the automated reboots and reply on a human to reboot it
during quieter not rebooty times