Disable traffic replay to Integration
...during data replication.

We are trying to reduce unnecessary Sentry errors, and one of those
ways is to disable traffic replay whilst the data sync is happening.

We manage traffic replay with our puppet class `govuk_gor`, setting
the Puppet `data_sync_in_progress` fact to `true` ensures that Puppet
does not try to run it until the data sync has completed.