Avoid referring to nonexistent Licensify TLS certs on AWS.
This should fix Puppet failures on `licensing-frontend` VMs which were
caused by the nginx config for the `uploadlicence` app referring to a
TLS certificate file which isn't there.

In AWS, we terminate TLS at the load balancers (ALBs) rather than at
nginx on the app VMs. Most of the config already handles this (via
`nginx_enable_ssl` in hiera) except for this piece of custom nginx
config in `licensing-frontend`.