Check the application memory twice before the check fails hard
Icinga has a distinction between soft and hard states for
checks. Setting the attempts_before_hard_state value to 2 (rather than
the default, 1) will mean Icinga won't notify about memory issues
until it's rechecked the memory for the app.

This'll hopefully avoid spurious notifications where the application
is being deployed, and naturally uses more memory for a short amount
of time.

This change was specifically motivated by seeing notifications for
this check for the Signon app.