Monitor application restarts due to memory usage
Icinga will restart applications when they exceed the warning memory
usage. This has the potential to hide issues with application memory
usage, as the application can be repeatedly restarted.

If an application is being repeatedly restarted, it could be that the
memory usage profile has changed, and either that needs investigating
or the check threshold needs adjusting accordingly.

This commit adds a new check to spot when this is happening, by
looking at the number of times an application has been restarted on
the current day.

I've been looking at this as the Publishing API looks to be in this
situation currently, and is being restarted ~5 times a day (across all
machines).