Increase the Publishing API memory check thresholds
The number of Unicorn workers was increased in [1], which looks to
have led to more memory being used (which seems to make
sense). Currently, the application is frequently being restarted by
Icinga, as the memory usage often exceeds the warning threshold.

To account for the increased memory usage, increase the
thresholds. Due to the application being frequently restarted, I'm not
sure what this should be set to, so just increase it a little bit.

1: c22812716cbdecd53367cf544d05dd3017d915ef