Increase ClamAV size limits to match upload size limits.
As of b7a26a2, we increased the max upload size in nginx from 30 to 100
MB. This increases clamd's size limits accordingly.

The 5 MB margin is to account for the ambiguity as to whether nginx's
"M" suffix refers to powers of 2 or powers of 10. (I wasn't able to find
an authoritative answer to this.)