Increase licensify 5xx rate in staging
We're seeing a problem where the 5xx alert often triggers in staging. We
think this is related to the app not handling errors properly.

Unfortunately it's unlikely we'll be able to prioritise work in this
area any time soon so I thought it was best to increase the thresholds
to stop it alerting.