Use ES recovery API to check snapshot restore is in progress
Concurrent snapshot restores aren't allowed, attempting to start a
restore while one is already happening will fail.  So previously we
only ran the restore on a machine in availability zone A.  This worked
fine until we shrunk down to two instances, which by chance are in
zones B and C.  A better solution is to trigger the restores, ignoring
errors, and use the recovery API to check that one of them has begun.

Making the second `curl` request immediately after the first fails a
lot, so also add in a small delay.