Add Jenkins job for deploying the developer docs
The job to build and deploy the developer docs is currently managed
manually in Jenkins. This is of course bad, but I never got around to
doing it the right way.

Managing it in puppet is better because it's versioned, we can keep the
secrets outside of the logs, and we get Icinga alerts when the build
fails.

Before this is merged I'll have to add the GitHub/AWS credentials to
govuk-secrets.