Increase the licensify 5xx thresholds in Integration
Same reason as d132947d0cd80d932249f5398a1aaa1241899e61, we're seeing
a problem where the 5xx alert often triggers. We think this is related
to the app not handling errors properly.

Unfortunately it's unlikely we'll be able to prioritise work in this
area any time soon so increase the thresholds to stop it alerting.