Stop including govuk_aws_xray_daemon on nodes
As we stopped using xray a while ago, and Puppet should have done the
cleanup now.

I'm looking at this as the Upstart service doesn't seem to get removed
in the cleanup, and as the Puppet config says to stop it, this breaks
govuk-puppet on new machines without the stopped xray service.