govuk_env_sync support for documentdb
mongodb is being replaced by documentdb in AWS so govuk_env_sync show have support for it

While we lay the configuration to pull licensify from staging daily is there, this is
not yet activated because the data is contains some incompatible record.
They have been tested to work in integration.

Further testing for licensify-audit is required as it sometimes fail if an app
recreates a collection which the govuk_env_sync is trying to do so too.