Remove pgbouncer code
The pgbouncer class was removed from db_admin nodes in #9236, but the
logic which sets up the pgbouncer configuration was not removed.  So
puppet is failing to run on any db_admin machines created after #9236
was merged: the pgbouncer files don't get created, but things still
reference them.

This is safe to remove because #9226 made all the apps talk to RDS
directly, not through pgbouncer.