Add fqdn variable sent to jenkins as a parameter
When a new server comes up, it will run the govuk_sync_apps
bash script after being first puppetized. This script
used to send a message to Jenkins to deploy all the apps for this
node class to all servers of this node class.

This is not ideal when multiple servers of the same class are
coming and some servers are not ready for provisioning by Jenkins.
Therefore, we add another parameter "TARGET_MACHINES" to the
message to Jenkins which provides the FQDN of the sending server
so that now Jenkins provisions only this machine.

The "Deploy_node_apps" is modified to take this new parameter
and pass it to the modified app deployment job "deploy_apps".