Move TARGET_MACHINES under TAG
It used to be the case that TAG was always under DEPLOY_TASK and since
we've added the new TARGET_MACHINES field I've entered the branch I
wanted to deploy into that field multiple times. Although this is just
memory that I could get used to, I think it also makes sense to put the
fields which are more likely to be used nearer the top. TARGET_MACHINES
is very unlikely to be manually changed by a human when deploying an
app whereas it's very common to change the TAG.