Revert "Update healthcheck path for Government Frontend"
Switch back to using /healthcheck, as using a item of content doesn't
seem appropriate for a healthcheck. I'm unsure what the behaviour of
Icinga, and the AWS load balancers will be if this item of content is
moved (resulting in a redirect), or unpublished (maybe resulting in a
gone response).

I think the motivation at the time was to have the healthcheck cover
more of the applications functionality, but the way to do that is
improve the healthcheck to cover more of the health of the
application.

This reverts commit dc67b8bfe3eabfc5a2a4ed470266d043b6532b5e.