Reduce time period for memory restarts alert
This was increased to 2days because it was causing Graphite
to return an error. This won't happen now, since I guess transformNull
is being used?

This won't make alerts go away if restarts are still happening, but will
prevent alerts from the previous day continuing to show up in Icinga.