Rate limit /contact/govuk without trailing slash
Currently the regular expression doesn't match against `/contact/govuk`
because it requires a trailing slash. By introducing an optional group
with a `/` at the beginning, we can make sure to match against both
`/contact/govuk` and `/contact/govuk/`.

I'm not very familiar with regular expressions - there might be a better
way of expressing this condition.