add option for puppet not to manage clamav running state
In some situations, clamav should only be set to run when its functionality
is needed. E.g. runninng licensify integration tests on the CI box.
Hence, we need to add an option for puppet to install clamav on the CI box
but not to manage its running state which will be managed by the
tasks which need clamav.