Disable the procfile worker memory check
The intent of adding the check was to get Icinga to restart the
service when lots of memory was being used, therefore avoiding failing
checks. However, the event handler doesn't currently work, as the
service sitting above the procfile workers doesn't support the reload
action.

Therefore, disable this check until the event handler can be fixed.