Don't replay Licensify traffic which assumes transactional state.
Certain GET requests to Licensify (that is, certain paths under
www.gov.uk/apply-for-a-licence/) only make sense after the license
application record has been created in the database. This happens via a
POST request which we intentionally exclude from traffic replay.

It therefore doesn't make sense to replay these requests - it isn't a
meaningful way to test those paths. The replayed requests cause errors
(because the expected state is missing from the database) and clutter up
the Staging logs with stacktraces, which frustrates debugging and could
mask other problems.