Create and include an lb_healthchecks.conf nginx config on all AWS boxes.
Currently there is a blob of copypasta on 7 node classes which use
lb_healthchecks.conf, but it needs to be everywhere on AWS because of
https://github.com/alphagov/govuk-aws/blob/master/doc/architecture/decisions/0037-alb-health-checks.md

This factors out the copypasta and paves the way for us to fix the LB
healthcheck config for other apps on AWS which are not currently
configured properly, include whitehall_backend/whitehall-admin.

With this change, an app should only need to pass $health_check_path to
govuk::app in order for its LB healthcheck to be mapped correctly on the
default vhost.