Remove service discovery override for Licensify in AWS Staging.
Licensify Staging is now running in AWS instead of UKCloud. This means
that apps running in AWS Staging must use the govuk-internal domain name
to connect to Licensify, not the external (public-facing) name.

We achieve this by removing the override of PLEK_SERVICE_LICENSIFY_URI
for apps in AWS Staging, so that Plek defaults to using the internal
name for app-to-app communication.