Disable Icinga flap detection by default
This is an Icinga feature to suppress notifications when the state of
a service or host changes frequently.

As we're using notifications now to try and track Icinga states in
Graphite, missing the notifications due to a service flapping means
that the information in Graphite is often out of date. Disabling flap
detection should make the situation better.

We're also using notifications for Slack and PagerDuty. I don't expect
this change to impact PagerDuty, as flap detection doesn't really
apply to the few services that notify PagerDuty, for example for the
5xx rate on the cache machines, the data is being smoothed out through
Graphite. As for Slack, this might increase the number of messages,
but it'll be more representative, which should be more useful.