Change govuk_nodes_spec_optional to test against AWS and Carrenza
Previously, the node specs would only be run against the Carrenza
hieradata, and some AWS specific node classes were excluded.

This commit changes the running, and the spec file to handle working
with both Carrenza and now AWS. I tried to handle both within the same
spec file, but due to hieradata related issues, the runner passes the
hosting provider in as an environment variable.

The parallel processing code had to be modified to account for this,
and while the change here is a little crude, just running the Carrenza
specs after all the AWS specs finish, it does work.

The lists of excluded classes should reflect which classes are
actually used with each hosting provider. The lists at the moment are
a mixture of classes I believe aren't applicable to that hosting
provider, and ones where the test fails, so there could be some
mistakes.

There are a couple of motivations behind these changes, firstly to
provide better testing coverage for AWS related Puppet configuration,
and secondly to allow the Carrenza hieradata to not have to support
AWS node classes.