Copy the production hieradata for the Local Links Manager to AWS
This configuration was added to Carrenza, but I assume the service
reverted to using the defaults (false in both cases) when it was
migrated to AWS.

I'm assuming both of these two things are still relevant. I can't see
any evidence of them being intentionally turned off, so we should
probably just turn them back on.