Remove Icinga count metric
It's currently unclear exactly what the value of this metric is because
the data coming back to Graphite isn't an integer count. It seems to end
up going in steps of 0.2 and we're not sure why.

Although this is possibly a configuration issue with Graphite or Icinga,
we don't actually need this metric anyway because all the information
you need can be calculated from the gauge metric we recently added.

I also think it doesn't make sense to use an `increment` here because
this script is only called on a state change, so it becomes slightly
misleading as it's not a count of the number of alerts, and more of a
count on the number of times an alert changes state (which, although
useful, might not be what people first expect).