Run 1, rather than 2 Sidekiq processes for Whitehall
I think this is more consistent with the behaviour for other apps. It
used to be that thread based concurrency wasn't used for Whitehall,
and instead many processes were used. Now we're using both thread and
process based concurrency, running 2 Sidekiq processes on each
machine, each using some number of threads.

To make the Whitehall behaviour more consistent with other apps,
switch to running one Sidekiq process per machine.

This will probalby reduce the potential throughput, but that can be
increased if necessary by increasing the number of threads.

The hieradata_aws change here is just to keep the hieradata
consistent, Whitehall isn't running in AWS yet, so it won't have any
affect.